+++ examples/blit.rs
@@ -2,9 +2,6 @@ extern crate doryen_rs;
 
 use doryen_rs::{App, AppOptions, Console, DoryenApi, Engine, TextAlign, UpdateEvent};
 
-const CONSOLE_WIDTH: u32 = 80;
-const CONSOLE_HEIGHT: u32 = 45;
-
 struct MyRoguelike {
     c1_pos: (i32, i32),
     c1_spd: (i32, i32),
@@ -33,18 +30,12 @@ fn move_con(pos: &mut (i32, i32), spd: &mut (i32, i32), size: (i32, i32)) {
 
 impl Engine for MyRoguelike {
     fn init(&mut self, _api: &mut dyn DoryenApi) {}
-    fn update(&mut self, _api: &mut dyn DoryenApi) -> Option<UpdateEvent> {
+    fn update(&mut self, api: &mut dyn DoryenApi) -> Option<UpdateEvent> {
         if self.step == 0 {
-            move_con(
-                &mut self.c1_pos,
-                &mut self.c1_spd,
-                (CONSOLE_WIDTH as i32, CONSOLE_HEIGHT as i32),
-            );
-            move_con(
-                &mut self.c2_pos,
-                &mut self.c2_spd,
-                (CONSOLE_WIDTH as i32, CONSOLE_HEIGHT as i32),
-            );
+            let con = api.con();
+            let size = (con.get_width() as i32, con.get_height() as i32);
+            move_con(&mut self.c1_pos, &mut self.c1_spd, size);
+            move_con(&mut self.c2_pos, &mut self.c2_spd, size);
         }
         self.alpha = (self.alpha + 0.01) % 1.0;
         self.step = (self.step + 1) % 10;
@@ -53,8 +44,8 @@ impl Engine for MyRoguelike {
     fn render(&mut self, api: &mut dyn DoryenApi) {
         let con = api.con();
         con.clear(Some((0, 0, 0, 255)), None, Some(' ' as u16));
-        for x in 0..CONSOLE_WIDTH as i32 {
-            for y in 0..CONSOLE_HEIGHT as i32 {
+        for x in 0..con.get_width() as i32 {
+            for y in 0..con.get_height() as i32 {
                 con.back(
                     x,
                     y,
@@ -67,8 +58,8 @@ impl Engine for MyRoguelike {
             }
         }
         con.print(
-            (CONSOLE_WIDTH / 2) as i32,
-            (CONSOLE_HEIGHT / 2) as i32,
+            (con.get_width() / 2) as i32,
+            (con.get_height() / 2) as i32,
             "You create offscreen consoles\nand blit them on other consoles",
             TextAlign::Center,
             Some((255, 255, 255, 255)),
@@ -130,17 +121,8 @@ impl MyRoguelike {
 
 fn main() {
     let mut app = App::new(AppOptions {
-        console_width: CONSOLE_WIDTH,
-        console_height: CONSOLE_HEIGHT,
-        screen_width: CONSOLE_WIDTH * 8,
-        screen_height: CONSOLE_HEIGHT * 8,
         window_title: "blitting demo".to_owned(),
-        font_path: "terminal_8x8.png".to_owned(),
-        vsync: true,
-        fullscreen: false,
-        show_cursor: true,
-        resizable: true,
-        intercept_close_request: false,
+        ..Default::default()
     });
     app.set_engine(Box::new(MyRoguelike::new()));
     app.run();