+++ examples/exit.rs
@@ -2,8 +2,6 @@ extern crate doryen_rs;
 
 use doryen_rs::{App, AppOptions, Color, DoryenApi, Engine, TextAlign, UpdateEvent};
 
-const CONSOLE_WIDTH: u32 = 80;
-const CONSOLE_HEIGHT: u32 = 45;
 const WHITE: Color = (255, 255, 255, 255);
 
 struct MyRoguelike {
@@ -61,17 +59,8 @@ impl MyRoguelike {
 
 fn main() {
     let mut app = App::new(AppOptions {
-        console_width: CONSOLE_WIDTH,
-        console_height: CONSOLE_HEIGHT,
-        screen_width: CONSOLE_WIDTH * 8,
-        screen_height: CONSOLE_HEIGHT * 8,
-        window_title: "my roguelike".to_owned(),
-        font_path: "terminal_8x8.png".to_owned(),
-        vsync: true,
-        fullscreen: false,
-        show_cursor: true,
-        resizable: true,
         intercept_close_request: true,
+        ..Default::default()
     });
     app.set_engine(Box::new(MyRoguelike::new()));
     app.run();