+++ examples/image.rs
@@ -2,9 +2,6 @@ extern crate doryen_rs;
 
 use doryen_rs::{App, AppOptions, DoryenApi, Engine, Image, UpdateEvent};
 
-const CONSOLE_WIDTH: u32 = 80;
-const CONSOLE_HEIGHT: u32 = 45;
-
 struct MyRoguelike {
     skull: Image,
     angle: f32,
@@ -24,8 +21,8 @@ impl Engine for MyRoguelike {
         con.clear(None, Some((0, 0, 0, 255)), None);
         self.skull.blit_ex(
             con,
-            (CONSOLE_WIDTH / 2) as f32,
-            (CONSOLE_HEIGHT / 2) as f32,
+            (con.get_width() / 2) as f32,
+            (con.get_height() / 2) as f32,
             scale,
             scale,
             self.angle,
@@ -47,17 +44,8 @@ impl MyRoguelike {
 
 fn main() {
     let mut app = App::new(AppOptions {
-        console_width: CONSOLE_WIDTH,
-        console_height: CONSOLE_HEIGHT,
-        screen_width: CONSOLE_WIDTH * 8,
-        screen_height: CONSOLE_HEIGHT * 8,
         window_title: "doryen-rs image demo".to_owned(),
-        font_path: "terminal_8x8.png".to_owned(),
-        vsync: true,
-        fullscreen: false,
-        show_cursor: true,
-        resizable: true,
-        intercept_close_request: false,
+        ..Default::default()
     });
     app.set_engine(Box::new(MyRoguelike::new()));
     app.run();