+++ examples/perf.rs
@@ -2,9 +2,6 @@ extern crate doryen_rs;
 
 use doryen_rs::{App, AppOptions, DoryenApi, Engine, TextAlign, UpdateEvent};
 
-const CONSOLE_WIDTH: u32 = 80;
-const CONSOLE_HEIGHT: u32 = 45;
-
 struct PerfTest {
     seed: u64,
 }
@@ -82,17 +79,9 @@ impl PerfTest {
 
 fn main() {
     let mut app = App::new(AppOptions {
-        console_width: CONSOLE_WIDTH,
-        console_height: CONSOLE_HEIGHT,
-        screen_width: CONSOLE_WIDTH * 8,
-        screen_height: CONSOLE_HEIGHT * 8,
         window_title: "doryen-rs performance test".to_owned(),
-        font_path: "terminal_8x8.png".to_owned(),
         vsync: false,
-        fullscreen: false,
-        show_cursor: true,
-        resizable: true,
-        intercept_close_request: false,
+        ..Default::default()
     });
     app.set_engine(Box::new(PerfTest::new()));
     app.run();