+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+# 1.2.1 - TBD
+## features
+* `AppOptions` now implements `Default`
+
 # 1.2.0 - 2019 Nov 22
 ## fixed
 * fix #13 Console.print_color with text containing ']'
 ## API breaks
-* added `AppOption.intercept_close_request` to intercept clicks on the window close button (native only). See 'exit' example
+* added `AppOptions.intercept_close_request` to intercept clicks on the window close button (native only). See 'exit' example
 ## features
 * added `Console.text_color_len`